Improving run toggling for UntilNoInput mode
Previously if you pressed run, then started walking, you wouldn't be running.
This changes that so that you'll be running as long as you're holding the run
button but then you can release it once you start moving and the running will
toggle off when you let go of input. Definitely feels nicer.